+++ src/librdag/equals.cc
@@ -18,7 +18,7 @@ namespace librdag {
 const real8 FuzzyEquals_default_maxabserror = std::numeric_limits<real8>::epsilon();
 const real8 FuzzyEquals_default_maxrelerror = 10*std::numeric_limits<real8>::epsilon();
 
-template <typename T> bool ArrayBitEquals(T * arr1, T * arr2, int count)
+template <typename T> bool ArrayBitEquals(T * arr1, T * arr2, size_t count)
 {
   assert(count>=0);
   size_t len = count*sizeof(T);
@@ -29,14 +29,14 @@ template <typename T> bool ArrayBitEquals(T * arr1, T * arr2, int count)
   return false;
 }
 
-template bool ArrayBitEquals(real8 * arr1, real8 * arr2, int count);
-template bool ArrayBitEquals(complex16 * arr1, complex16 * arr2, int count);
-template bool ArrayBitEquals(int * arr1, int * arr2, int count);
+template bool ArrayBitEquals(real8 * arr1, real8 * arr2, size_t count);
+template bool ArrayBitEquals(complex16 * arr1, complex16 * arr2, size_t count);
+template bool ArrayBitEquals(int4 * arr1, int4 * arr2, size_t count);
 
-template <typename T> bool ArrayFuzzyEquals(T * arr1, T * arr2, int count, real8 maxabserror, real8 maxrelerror)
+template <typename T> bool ArrayFuzzyEquals(T * arr1, T * arr2, size_t count, real8 maxabserror, real8 maxrelerror)
 {
   assert(count>=0);
-  for(int i=0;i<count;i++)
+  for(size_t i=0;i<count;i++)
   {
     if(!SingleValueFuzzyEquals(arr1[i], arr2[i], maxabserror, maxrelerror)) return false;
   }
@@ -44,8 +44,8 @@ template <typename T> bool ArrayFuzzyEquals(T * arr1, T * arr2, int count, real8
   return true;
 }
 
-template bool ArrayFuzzyEquals(real8 * arr1, real8 * arr2, int count, real8 maxabserror, real8 maxrelerror);
-template bool ArrayFuzzyEquals(complex16 * arr1, complex16 * arr2, int count, real8 maxabserror, real8 maxrelerror);
+template bool ArrayFuzzyEquals(real8 * arr1, real8 * arr2, size_t count, real8 maxabserror, real8 maxrelerror);
+template bool ArrayFuzzyEquals(complex16 * arr1, complex16 * arr2, size_t count, real8 maxabserror, real8 maxrelerror);
 
 
 /**
